# v2/server/pkg/contracts/openapi_gen.go
// Package contracts provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package contracts

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for ActionRequestRisk.
const (
	High   ActionRequestRisk = "high"
	Low    ActionRequestRisk = "low"
	Medium ActionRequestRisk = "medium"
)

// Defines values for ActionResultDecision.
const (
	Allow ActionResultDecision = "allow"
	Ask   ActionResultDecision = "ask"
	Deny  ActionResultDecision = "deny"
)

// Defines values for ManagedSessionStatus.
const (
	AwaitingApproval ManagedSessionStatus = "awaiting_approval"
	Cancelled        ManagedSessionStatus = "cancelled"
	Completed        ManagedSessionStatus = "completed"
	Failed           ManagedSessionStatus = "failed"
	Pending          ManagedSessionStatus = "pending"
	Running          ManagedSessionStatus = "running"
)

// ActionRequest defines model for ActionRequest.
type ActionRequest struct {
	// CapabilityToken Server-signed capability envelope; verified at use-time.
	CapabilityToken   *string                 `json:"capabilityToken,omitempty"`
	Context           *map[string]interface{} `json:"context,omitempty"`
	DomainAllowListed *bool                   `json:"domainAllowListed,omitempty"`
	InScope           *bool                   `json:"inScope,omitempty"`
	OutsideScope      *bool                   `json:"outsideScope,omitempty"`

	// Resource Resource class, e.g. fs.read, fs.write, net, shell, data_export, tool:<name>.
	Resource string             `json:"resource"`
	Risk     *ActionRequestRisk `json:"risk,omitempty"`
	TaskId   string             `json:"taskId"`
}

// ActionRequestRisk defines model for ActionRequest.Risk.
type ActionRequestRisk string

// ActionResult defines model for ActionResult.
type ActionResult struct {
	Decision ActionResultDecision `json:"decision"`
	Events   *[]CommittedEvent    `json:"events,omitempty"`

	// RuleId Matched permission rule id (audit).
	RuleId string `json:"ruleId"`
}

// ActionResultDecision defines model for ActionResult.Decision.
type ActionResultDecision string

// Budget defines model for Budget.
type Budget struct {
	MaxCostUsd      *float32 `json:"maxCostUsd,omitempty"`
	MaxTokens       *int     `json:"maxTokens,omitempty"`
	MaxWallClockSec *int     `json:"maxWallClockSec,omitempty"`
}

// Command defines model for Command.
type Command struct {
	// ExpectedStreamSeq Optimistic concurrency guard; reject if the stream advanced.
	ExpectedStreamSeq *int64 `json:"expectedStreamSeq,omitempty"`

	// IdempotencyKey Dedup key for safe client retries.
	IdempotencyKey *string `json:"idempotencyKey,omitempty"`

	// Payload Command-specific fields (validated against the command schema).
	Payload map[string]interface{} `json:"payload"`

	// Type Command name, e.g. CreateTask, StartTurn, ResolveApproval.
	Type string `json:"type"`
}

// CommittedEvent Event envelope as appended to event_log. The payload conforms to the JSON Schema named by `type` under contracts/events/.
type CommittedEvent struct {
	Actor         string                 `json:"actor"`
	CausationId   *string                `json:"causationId,omitempty"`
	CorrelationId *string                `json:"correlationId,omitempty"`
	GlobalSeq     int64                  `json:"globalSeq"`
	OccurredAt    time.Time              `json:"occurredAt"`
	Payload       map[string]interface{} `json:"payload"`
	SchemaVer     int                    `json:"schemaVer"`

	// StreamId Aggregate instance, e.g. task:<id>
	StreamId  string `json:"streamId"`
	StreamSeq int64  `json:"streamSeq"`
	TenantId  string `json:"tenantId"`
	Type      string `json:"type"`
}

// Cursor Opaque resume position; encodes the per-shard monotonic visible global_seq.
type Cursor = string

// DomainError Typed, fail-fast error. One handling site per error class.
type DomainError struct {
	// Code Stable machine code, e.g. concurrency_conflict, invariant_violated, capability_denied.
	Code          string                  `json:"code"`
	CorrelationId *string                 `json:"correlationId,omitempty"`
	Details       *map[string]interface{} `json:"details,omitempty"`
	Message       string                  `json:"message"`
}

// ManagedSession defines model for ManagedSession.
type ManagedSession struct {
	Id     string                  `json:"id"`
	Result *map[string]interface{} `json:"result,omitempty"`
	Status ManagedSessionStatus    `json:"status"`
	TaskId *string                 `json:"taskId,omitempty"`
}

// ManagedSessionStatus defines model for ManagedSession.Status.
type ManagedSessionStatus string

// ReadModelPage defines model for ReadModelPage.
type ReadModelPage struct {
	Items []map[string]interface{} `json:"items"`

	// NextCursor Opaque resume position; encodes the per-shard monotonic visible global_seq.
	NextCursor *Cursor `json:"nextCursor,omitempty"`
}

// RunQueryParams defines parameters for RunQuery.
type RunQueryParams struct {
	Cursor *Cursor `form:"cursor,omitempty" json:"cursor,omitempty"`
	Limit  *int    `form:"limit,omitempty" json:"limit,omitempty"`
}

// CreateManagedSessionJSONBody defines parameters for CreateManagedSession.
type CreateManagedSessionJSONBody struct {
	Budget      *Budget `json:"budget,omitempty"`
	Prompt      string  `json:"prompt"`
	WorkspaceId *string `json:"workspaceId,omitempty"`
}

// SubscribeStreamParams defines parameters for SubscribeStream.
type SubscribeStreamParams struct {
	Cursor *Cursor `form:"cursor,omitempty" json:"cursor,omitempty"`
}

// AuthorizeActionJSONRequestBody defines body for AuthorizeAction for application/json ContentType.
type AuthorizeActionJSONRequestBody = ActionRequest

// DispatchCommandJSONRequestBody defines body for DispatchCommand for application/json ContentType.
type DispatchCommandJSONRequestBody = Command

// CreateManagedSessionJSONRequestBody defines body for CreateManagedSession for application/json ContentType.
type CreateManagedSessionJSONRequestBody CreateManagedSessionJSONBody